Redesigning settings layout & adding in support for checking head tracking enablement
