+++ EyeSpeak/Common/Views/GazeableButton.swift
@@ -9,7 +9,9 @@
 import Foundation
 import UIKit
 
-class VocableUIControl: UIControl {
+class GazeableButton: UIButton {
+    
+    var beginDate = Date()
     
     let borderedView = BorderedView()
     
@@ -39,9 +41,18 @@ class VocableUIControl: UIControl {
         borderedView.cornerRadius = 8
         borderedView.borderColor = .cellBorderHighlightColor
         borderedView.backgroundColor = .collectionViewBackgroundColor
-        
+        borderedView.isUserInteractionEnabled = false
+        borderedView.fillColor = .clear
+
         updateContentViews()
         addSubview(borderedView)
+        borderedView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            borderedView.topAnchor.constraint(equalTo: self.topAnchor, constant: 0),
+            borderedView.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 0),
+            borderedView.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: 0),
+            borderedView.bottomAnchor.constraint(equalTo: self.bottomAnchor, constant: 0)
+        ])
     }
     
     func updateContentViews() {
@@ -50,16 +61,31 @@ class VocableUIControl: UIControl {
         borderedView.isOpaque = true
     }
     
+    override var canReceiveGaze: Bool {
+        return true
+    }
+    
     override func gazeBegan(_ gaze: UIHeadGaze, with event: UIHeadGazeEvent?) {
-        self.defaultBackgroundColor = .green
+        super.gazeBegan(gaze, with: event)
+        
+        isHighlighted = true
+        beginDate = Date()
     }
 
     override func gazeMoved(_ gaze: UIHeadGaze, with event: UIHeadGazeEvent?) {
-        self.defaultBackgroundColor = .blue
+        super.gazeBegan(gaze, with: event)
+        
+        let timeElapsed = Date().timeIntervalSince(beginDate)
+        if timeElapsed >= gaze.selectionHoldDuration {
+            isSelected = true
+        }
     }
 
     override func gazeEnded(_ gaze: UIHeadGaze, with event: UIHeadGazeEvent?) {
-        self.defaultBackgroundColor = .red
+        super.gazeBegan(gaze, with: event)
+        
+        isSelected = false
+        isHighlighted = false
     }
     
 }
