+++ EyeSpeak/Features/Settings/SettingsToggleCollectionViewCell.xib
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
+        <capability name="Named colors" minToolsVersion="9.0"/>
+        <capability name="collection view cell content view" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="nEO-5f-YSR" customClass="SettingsToggleCollectionViewCell" customModule="EyeSpeak" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="523" height="99"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="0GQ-x3-4PV">
+                <rect key="frame" x="0.0" y="0.0" width="523" height="99"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="6VK-zy-47F">
+                        <rect key="frame" x="16" y="36" width="53" height="27"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="22"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <switch opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" contentHorizontalAlignment="center" contentVerticalAlignment="center" on="YES" translatesAutoresizingMaskIntoConstraints="NO" id="rux-wr-INz">
+                        <rect key="frame" x="458" y="34" width="51" height="31"/>
+                    </switch>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="KXC-cg-sCS">
+                        <rect key="frame" x="0.0" y="0.0" width="523" height="2"/>
+                        <color key="backgroundColor" name="DefaultCellBackground"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="2" id="zov-hU-Tw8"/>
+                        </constraints>
+                    </view>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="S25-62-sbz">
+                        <rect key="frame" x="0.0" y="97" width="523" height="2"/>
+                        <color key="backgroundColor" name="DefaultCellBackground"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="2" id="DXX-dP-Pwd"/>
+                        </constraints>
+                    </view>
+                </subviews>
+                <constraints>
+                    <constraint firstAttribute="trailing" secondItem="KXC-cg-sCS" secondAttribute="trailing" id="58n-wi-I7W"/>
+                    <constraint firstItem="KXC-cg-sCS" firstAttribute="leading" secondItem="0GQ-x3-4PV" secondAttribute="leading" id="6rL-CK-DOf"/>
+                    <constraint firstItem="6VK-zy-47F" firstAttribute="leading" secondItem="0GQ-x3-4PV" secondAttribute="leading" constant="16" id="Acm-dF-bue"/>
+                    <constraint firstAttribute="trailing" secondItem="rux-wr-INz" secondAttribute="trailing" constant="16" id="C1k-Fw-WXx"/>
+                    <constraint firstAttribute="trailing" secondItem="S25-62-sbz" secondAttribute="trailing" id="J1G-Jl-T1h"/>
+                    <constraint firstItem="KXC-cg-sCS" firstAttribute="top" secondItem="0GQ-x3-4PV" secondAttribute="top" id="Job-43-ylb"/>
+                    <constraint firstItem="rux-wr-INz" firstAttribute="centerY" secondItem="0GQ-x3-4PV" secondAttribute="centerY" id="P1e-pV-gFd"/>
+                    <constraint firstItem="6VK-zy-47F" firstAttribute="centerY" secondItem="0GQ-x3-4PV" secondAttribute="centerY" id="ZkI-fu-JDe"/>
+                    <constraint firstItem="S25-62-sbz" firstAttribute="leading" secondItem="0GQ-x3-4PV" secondAttribute="leading" id="hvW-zg-uCM"/>
+                    <constraint firstAttribute="bottom" secondItem="S25-62-sbz" secondAttribute="bottom" id="l4H-FN-Aex"/>
+                </constraints>
+            </collectionViewCellContentView>
+            <size key="customSize" width="523" height="99"/>
+            <connections>
+                <outlet property="enabledSwitch" destination="rux-wr-INz" id="AVi-4f-1Ui"/>
+                <outlet property="textLabel" destination="6VK-zy-47F" id="niV-LV-CDy"/>
+            </connections>
+            <point key="canvasLocation" x="-667.39130434782612" y="-405.46875"/>
+        </collectionViewCell>
+    </objects>
+    <resources>
+        <namedColor name="DefaultCellBackground">
+            <color red="0.21799999475479126" green="0.19499999284744263" blue="0.62699997425079346" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+    </resources>
+</document>
